requirements.txt
fastapi
uvicorn
openai
python-dotenv
httpx==0.27.2
groq